refactor: mv findTarget to inspectionUtils
