+++ packages/broker/test/unit/plugins/operator/inspectionUtils.test.ts
@@ -2,10 +2,14 @@ import { FetchRedundancyFactorFn } from '../../../../src/plugins/operator/Inspec
 import { mock, MockProxy } from 'jest-mock-extended'
 import { MessageListener, NetworkPeerDescriptor, StreamrClient, Subscription } from 'streamr-client'
 import { createHeartbeatMessage } from '../../../../src/plugins/operator/heartbeatUtils'
-import { toStreamID, toStreamPartID } from '@streamr/protocol'
+import { StreamID, StreamPartID, toStreamID, toStreamPartID } from '@streamr/protocol'
 import { randomEthereumAddress } from '@streamr/test-utils'
-import { findNodesForTarget, inspectTarget } from '../../../../src/plugins/operator/inspectionUtils'
+import { findNodesForTarget, findTarget, inspectTarget } from '../../../../src/plugins/operator/inspectionUtils'
+import { InspectRandomNodeHelper } from '../../../../src/plugins/operator/InspectRandomNodeHelper'
+import { StreamAssignmentLoadBalancer } from '../../../../src/plugins/operator/StreamAssignmentLoadBalancer'
+import { EthereumAddress } from '@streamr/utils'
 
+const MY_OPERATOR_ADDRESS = randomEthereumAddress()
 const OTHER_OPERATOR_ADDRESS = randomEthereumAddress()
 const SPONSORSHIP_ADDRESS = randomEthereumAddress()
 const STREAM_ID = toStreamID('streamId')
@@ -19,6 +23,86 @@ const target = Object.freeze({
 const PEER_DESCRIPTOR_ONE = { id: '0x1111' }
 const PEER_DESCRIPTOR_TWO = { id: '0x2222' }
 const PEER_DESCRIPTOR_THREE = { id: '0x3333' }
+
+describe(findTarget, () => {
+    let helper: MockProxy<InspectRandomNodeHelper>
+    let loadBalancer: MockProxy<StreamAssignmentLoadBalancer>
+
+    function setupEnv(sponsorships: Array<{ address: EthereumAddress, operators: EthereumAddress[], streamId: StreamID }>) {
+        helper.getSponsorshipsOfOperator.mockImplementation(async (operatorAddress) => {
+            return sponsorships
+                .filter(({ operators }) => operators.includes(operatorAddress))
+                .map(({ address, operators, streamId }) => ({
+                    sponsorshipAddress: address,
+                    operatorCount: operators.length,
+                    streamId,
+                }))
+        })
+        helper.getOperatorsInSponsorship.mockImplementation(async (sponsorshipAddress) => {
+            return sponsorships.find(({ address }) => address === sponsorshipAddress)!.operators
+        })
+    }
+
+    function setStreamPartsAssignedToMe(streamParts: StreamPartID[]): void {
+        loadBalancer.getMyStreamParts.mockReturnValue(streamParts)
+    }
+
+    beforeEach(() => {
+        helper = mock<InspectRandomNodeHelper>()
+        loadBalancer = mock<StreamAssignmentLoadBalancer>()
+    })
+
+    it('returns undefined if no sponsorships are found', async () => {
+        setupEnv([])
+        setStreamPartsAssignedToMe([])
+        const result = await findTarget(MY_OPERATOR_ADDRESS, helper, loadBalancer)
+        expect(result).toBeUndefined()
+    })
+
+    it('returns undefined if only finds sponsorship with my operator as only operator', async () => {
+        setupEnv([{
+            address: SPONSORSHIP_ADDRESS,
+            operators: [MY_OPERATOR_ADDRESS],
+            streamId: STREAM_ID,
+        }])
+        setStreamPartsAssignedToMe([])
+        const result = await findTarget(MY_OPERATOR_ADDRESS, helper, loadBalancer)
+        expect(result).toBeUndefined()
+    })
+
+    it('returns undefined if no sponsorships found with a partition assigned to me', async () => {
+        setupEnv([{
+            address: SPONSORSHIP_ADDRESS,
+            operators: [MY_OPERATOR_ADDRESS, OTHER_OPERATOR_ADDRESS],
+            streamId: STREAM_ID,
+        }])
+        setStreamPartsAssignedToMe([])
+        const result = await findTarget(MY_OPERATOR_ADDRESS, helper, loadBalancer)
+        expect(result).toBeUndefined()
+    })
+
+    it('returns target sponsorship, operator and stream part', async () => {
+        setupEnv([{
+            address: SPONSORSHIP_ADDRESS,
+            operators: [MY_OPERATOR_ADDRESS, OTHER_OPERATOR_ADDRESS],
+            streamId: STREAM_ID,
+        }])
+        setStreamPartsAssignedToMe([
+            toStreamPartID(STREAM_ID, 0),
+            toStreamPartID(STREAM_ID, 1),
+            toStreamPartID(STREAM_ID, 2),
+        ])
+
+        const result = await findTarget(MY_OPERATOR_ADDRESS, helper, loadBalancer)
+        expect(result).toMatchObject({
+            sponsorshipAddress: SPONSORSHIP_ADDRESS,
+            operatorAddress: OTHER_OPERATOR_ADDRESS,
+            streamPart: expect.stringMatching(/^streamId#\d$/)
+        })
+    })
+
+    // TODO: few edge-cases where state changes during asynchronicity
+})
 
 describe(findNodesForTarget, () => {
     let streamrClient: MockProxy<StreamrClient>
